backend/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
sentence-transformers
faiss-cpu
google-generativeai
pydantic # FastAPI uses this for data validation
